refactor(processMarkdown): moved into a script that runs daily from cron
